+++ apps/web/src/App.tsx
@@ -20,6 +20,7 @@ import { PromptBar } from './ask/PromptBar';
 import { CardActionBar } from './ask/CardActionBar';
 import { ClarifyLayer } from './ask/ClarifyLayer';
 import { DraftControls } from './ask/DraftControls';
+import { ProfileOfferLayer } from './ask/ProfileOfferLayer';
 import { RegenControls } from './ask/RegenControls';
 import { SelectionAsk } from './ask/SelectionAsk';
 import { Timeline } from './log/Timeline';
@@ -57,6 +58,7 @@ function JarwizOverlay() {
       <EmptyState />
       <BoardEntry />
       <CardActionBar />
+      <ProfileOfferLayer />
       <ClarifyLayer />
       <DraftControls />
       <RegenControls />